refactor: move sendRegister to action
