+++ src/actions/sendRegister.ts
